// packages/storage-model-atlas/src/generated/workflowapi/Scope.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject } from '@emfts/core';
import type { Registry } from './Registry';

/**
 * Scope
 * @generated
 */
export interface Scope extends EObject {
  name?: string;
  description?: string;
  parentScope?: string;
  registries: Registry[];

}